onBeforeShapeIsolate?

+++ packages/editor/src/lib/editor/Editor.ts
@@ -109,8 +109,7 @@ import { getReorderingShapesChanges } from '../utils/reorderShapes'
 import { applyRotationToSnapshotShapes, getRotationSnapshot } from '../utils/rotation'
 import { uniqueId } from '../utils/uniqueId'
 import {
-	BindingOnUnbindOptions,
-	BindingUnbindReason,
+	BindingOnDeleteOptions,
 	BindingUtil,
 	TLBindingUtilConstructor,
 } from './bindings/BindingUtil'
@@ -355,7 +354,7 @@ export class Editor extends EventEmitter<TLEventMap> {
 
 		this.sideEffects = this.store.sideEffects
 
-		let deletedBindings = new Map<TLBindingId, BindingOnUnbindOptions<any>>()
+		let deletedBindings = new Map<TLBindingId, BindingOnDeleteOptions<any>>()
 		const deletedShapeIds = new Set<TLShapeId>()
 		const invalidParents = new Set<TLShapeId>()
 		let invalidBindingTypes = new Set<string>()
@@ -391,7 +390,7 @@ export class Editor extends EventEmitter<TLEventMap> {
 					const t = deletedBindings
 					deletedBindings = new Map()
 					for (const opts of t.values()) {
-						this.getBindingUtil(opts.binding).onAfterUnbind?.(opts)
+						this.getBindingUtil(opts.binding).onAfterDelete?.(opts)
 					}
 				}
 
@@ -487,8 +486,19 @@ export class Editor extends EventEmitter<TLEventMap> {
 						for (const binding of this.getBindingsInvolvingShape(shape)) {
 							invalidBindingTypes.add(binding.type)
 							deleteBindingIds.push(binding.id)
+							const util = this.getBindingUtil(binding)
+							if (binding.fromId === shape.id) {
+								util.onBeforeIsolateToShape?.({ binding, shape })
+								util.onBeforeDeleteFromShape?.({ binding, shape })
+							} else {
+								util.onBeforeIsolateFromShape?.({ binding, shape })
+								util.onBeforeDeleteToShape?.({ binding, shape })
+							}
 						}
-						this.deleteBindings(deleteBindingIds)
+
+						if (deleteBindingIds.length) {
+							this.deleteBindings(deleteBindingIds)
+						}
 
 						const deletedIds = new Set([shape.id])
 						const updates = compact(
@@ -525,25 +535,10 @@ export class Editor extends EventEmitter<TLEventMap> {
 						this.getBindingUtil(bindingAfter).onAfterChange?.({ bindingBefore, bindingAfter })
 					},
 					beforeDelete: (binding) => {
-						const util = this.getBindingUtil(binding)
-						// No need to track this binding if it's util doesn't care about the unbind operation
-						if (!util.onBeforeUnbind && !util.onAfterUnbind) return
-						// We only want to call this once per binding and it might be possible that the onBeforeUnbind
-						// callback will trigger a nested delete operation on the same binding so let's bail out if
-						// that is happening
-						if (deletedBindings.has(binding.id)) return
-						const opts: BindingOnUnbindOptions<any> = {
-							binding,
-							reason: deletedShapeIds.has(binding.fromId)
-								? BindingUnbindReason.DeletingFromShape
-								: deletedShapeIds.has(binding.toId)
-									? BindingUnbindReason.DeletingToShape
-									: BindingUnbindReason.DeletingBinding,
-						}
-						deletedBindings.set(binding.id, opts)
-						this.getBindingUtil(binding).onBeforeUnbind?.(opts)
+						this.getBindingUtil(binding).onBeforeDelete?.({ binding })
 					},
 					afterDelete: (binding) => {
+						this.getBindingUtil(binding).onAfterDelete?.({ binding })
 						invalidBindingTypes.add(binding.type)
 					},
 				},
@@ -3448,7 +3443,7 @@ export class Editor extends EventEmitter<TLEventMap> {
 	}
 	private _tickCameraState = () => {
 		// always reset the timeout
-		this._cameraStateTimeoutRemaining = this.options.cameraMovingTimoutMs
+		this._cameraStateTimeoutRemaining = this.options.cameraMovingTimeout
 		// If the state is idle, then start the tick
 		if (this._cameraState.__unsafe__getWithoutCapture() !== 'idle') return
 		this._cameraState.set('moving')
@@ -5234,13 +5229,22 @@ export class Editor extends EventEmitter<TLEventMap> {
 		return this.updateBindings([partial])
 	}
 
-	deleteBindings(bindings: (TLBinding | TLBindingId)[]) {
+	deleteBindings(bindings: (TLBinding | TLBindingId)[], { isolateShapes = false } = {}) {
 		const ids = bindings.map((binding) => (typeof binding === 'string' ? binding : binding.id))
+		if (isolateShapes) {
+			for (const id of ids) {
+				const binding = this.getBinding(id)
+				if (!binding) continue
+				const util = this.getBindingUtil(binding)
+				util.onBeforeIsolateFromShape?.({ binding, shape: this.getShape(binding.fromId)! })
+				util.onBeforeIsolateToShape?.({ binding, shape: this.getShape(binding.toId)! })
+			}
+		}
 		this.store.remove(ids)
 		return this
 	}
-	deleteBinding(binding: TLBinding | TLBindingId) {
-		return this.deleteBindings([binding])
+	deleteBinding(binding: TLBinding | TLBindingId, opts?: Parameters<this['deleteBindings']>[1]) {
+		return this.deleteBindings([binding], opts)
 	}
 	canBindShapes({
 		fromShape,
@@ -5393,7 +5397,7 @@ export class Editor extends EventEmitter<TLEventMap> {
 				shapeIds.set(shapeId, createShapeId())
 			}
 
-			const { shapesToCreate, bindingsToCreate } = withoutBindingsToUnrelatedShapes(
+			const { shapesToCreate, bindingsToCreate } = withIsolatedShapes(
 				this,
 				shapeIdSet,
 				(bindingIdsToMaintain) => {
@@ -7651,7 +7655,7 @@ export class Editor extends EventEmitter<TLEventMap> {
 
 		const shapeIds = this.getShapeAndDescendantIds(ids)
 
-		return withoutBindingsToUnrelatedShapes(this, shapeIds, (bindingIdsToKeep) => {
+		return withIsolatedShapes(this, shapeIds, (bindingIdsToKeep) => {
 			const bindings: TLBinding[] = []
 			for (const id of bindingIdsToKeep) {
 				const binding = this.getBinding(id)
@@ -8943,36 +8947,44 @@ function pushShapeWithDescendants(editor: Editor, id: TLShapeId, result: TLShape
  *
  * The callback is given the set of bindings that should be maintained.
  */
-function withoutBindingsToUnrelatedShapes<T>(
+function withIsolatedShapes<T>(
 	editor: Editor,
 	shapeIds: Set<TLShapeId>,
 	callback: (bindingsWithBoth: Set<TLBindingId>) => T
 ): T {
-	const bindingsWithBoth = new Set<TLBindingId>()
-	const bindingsToRemove = new Set<TLBindingId>()
-
-	for (const shapeId of shapeIds) {
-		const shape = editor.getShape(shapeId)
-		if (!shape) continue
-
-		for (const binding of editor.getBindingsInvolvingShape(shapeId)) {
-			const hasFrom = shapeIds.has(binding.fromId)
-			const hasTo = shapeIds.has(binding.toId)
-			if (hasFrom && hasTo) {
-				bindingsWithBoth.add(binding.id)
-				continue
-			}
-			if (!hasFrom || !hasTo) {
-				bindingsToRemove.add(binding.id)
-			}
-		}
-	}
-
 	let result!: Result<T, unknown>
 
 	editor.history.ignore(() => {
 		const changes = editor.store.extractingChanges(() => {
-			editor.deleteBindings([...bindingsToRemove])
+			const bindingsWithBoth = new Set<TLBindingId>()
+			const bindingsToRemove = new Set<TLBindingId>()
+
+			for (const shapeId of shapeIds) {
+				const shape = editor.getShape(shapeId)
+				if (!shape) continue
+
+				for (const binding of editor.getBindingsInvolvingShape(shapeId)) {
+					const hasFrom = shapeIds.has(binding.fromId)
+					const hasTo = shapeIds.has(binding.toId)
+					if (hasFrom && hasTo) {
+						bindingsWithBoth.add(binding.id)
+						continue
+					}
+					if (!hasFrom || !hasTo) {
+						bindingsToRemove.add(binding.id)
+					}
+				}
+			}
+
+			for (const bindingId of bindingsToRemove) {
+				const binding = editor.getBinding(bindingId)!
+				const util = editor.getBindingUtil(binding)
+				if (shapeIds.has(binding.fromId)) {
+					util.onBeforeIsolateFromShape?.({ binding, shape: editor.getShape(binding.fromId)! })
+				} else {
+					util.onBeforeIsolateToShape?.({ binding, shape: editor.getShape(binding.toId)! })
+				}
+			}
 
 			try {
 				result = Result.ok(callback(bindingsWithBoth))
